换了解析的思路

+++ src/adaptive_filter.cpp
@@ -1,10 +1,10 @@
 #include "../include/serial_connect/adaptive_filter.h"
 #include <cmath>
 
-AdaptiveFilter::AdaptiveFilter(int bufferSize, double threshold)
+AdaptiveFilter::AdaptiveFilter(int bufferSize, int threshold)
     : bufferSize(bufferSize), threshold(threshold) {}
 
-double AdaptiveFilter::filter(double value) {
+int AdaptiveFilter::filter(int value) {
     buffer.push_back(value);
 
     if (buffer.size() > bufferSize) {
@@ -12,25 +12,38 @@ double AdaptiveFilter::filter(double value) {
     }
 
     if (isOutlier(value)) {
-        double average = calculateAverage();
+        int average = calculateAverage();
         return average;
     } else {
         return value;
     }
 }
+// int AdaptiveFilter::filter(int value) {
+//     buffer.push_back(value);
 
-bool AdaptiveFilter::isOutlier(double value) const {
+//     if (buffer.size() > bufferSize) {
+//         buffer.pop_front();
+//     }
+
+//     if (isOutlier(value)) {
+//         int average = calculateAverage();
+//         return average;
+//     } else {
+//         return value;
+//     }
+// }
+bool AdaptiveFilter::isOutlier(int value) const {
     if (buffer.size() < 4) {
         return false;
     }
 
-    double lastValue = buffer[buffer.size() - 2];
-    double secondLastValue = buffer[buffer.size() - 3];
-    double thirdLastValue = buffer[buffer.size() - 4];
+    int lastValue = buffer[buffer.size() - 2];
+    int secondLastValue = buffer[buffer.size() - 3];
+    int thirdLastValue = buffer[buffer.size() - 4];
 
-    double diff = std::abs(value - lastValue);
-    double prevDiff = std::abs(lastValue - secondLastValue);
-    double secondDiff = std::abs(secondLastValue - thirdLastValue);
+    int diff = std::abs(value - lastValue);
+    int prevDiff = std::abs(lastValue - secondLastValue);
+    int secondDiff = std::abs(secondLastValue - thirdLastValue);
 
     std::cout<<"diff:"<<diff<<" "<<prevDiff<<std::endl;
     if (diff > threshold && diff > prevDiff *2 && diff > secondDiff *2 && value > 300) {
@@ -41,9 +54,9 @@ bool AdaptiveFilter::isOutlier(double value) const {
     return false;
 }
 
-double AdaptiveFilter::calculateAverage() const {
-    double sum = 0.0;
-    for (double value : buffer) {
+int AdaptiveFilter::calculateAverage() const {
+    int sum = 0.0;
+    for (int value : buffer) {
         sum += value;
     }
     return sum / buffer.size();
